Trait dex_manager_rpc::DEXManagerAPIServer

pub trait DEXManagerAPIServer<BlockHash, DEXId>: Sized + Send + Sync + 'static {
    fn list_dex_ids(&self, at: Option<BlockHash>) -> Result<Vec<DEXId>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        DEXId: Send + Sync + 'static + Serialize,
    { ... }
}

Server trait implementation for the DEXManagerAPI RPC API.

Required Methods

fn list_dex_ids(&self, at: Option<BlockHash>) -> Result<Vec<DEXId>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
DEXId: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.